requirements.txt
fastapi
uvicorn
prometheus_client
psycopg2-binary